pyproject.toml
[project]
name = "trading-agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Gajesh Naik",email = "26431906+Gajesh2007@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4"
dependencies = [
    "hyperliquid-python-sdk (>=0.20.0,<0.21.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "web3 (>=7.14.0,<8.0.0)",
    "aiohttp (>=3.13.1,<4.0.0)",
    "openai (>=2.5.0,<3.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "rich (>=14.2.0,<15.0.0)",
    "python-pptx (>=1.0.0,<2.0.0)",
    "chromadb (>=0.6.0,<1.0.0)",
    "Pillow (>=10.0.0,<12.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
